Ensure strict and culture-invariant date-time parsing in EventTimestamp Setter
This change applies CultureInfo.InvariantCulture and DateTimeStyles.None to the DateTime.Parse method in the EventTimestamp setter. 
It enforces strict parsing of the date-time string, avoiding potential issues with locale-specific formats and ensuring consistent behavior across different environments. Currently there is 
issue of when event_time is set on a foreign locale / computer it sets to a value of 2024-02-26T07.00.58.787372000Z instead of 2024-02-26T07:00:58.787372000Z and throws error becasue not in 
ISO 8601 format. This helps insure IDE will retrieve EventTimeStamp in consistent format.